refactor: Add typing

+++ python/needle/nn/softmax.py
@@ -3,8 +3,6 @@ otherwise they will overflow the division, thus making it a float64 result,
 which will cause type errors downstream.
 """
 
-from numpy import float32
-
 from needle import init, ops
 from needle.nn.nn_basic import Module
 from needle.tensor import Tensor
@@ -16,5 +14,4 @@ class SoftmaxLoss(Module):
         diff = (logits * y_one_hot).sum(axes=1)
 
         lse = ops.logsumexp(logits, axes=1)
-        # division causes result to be float64
-        return (lse - diff).sum() / float32(logits.shape[0])
+        return (lse - diff).sum() / logits.shape[0]
